nettest: refactor to allow returning async events
This patch changes the nettest model to allow a nettest to
return async events during its lifecycle.

This is currently not used by the Psiphon nettest and possibly not
needed soon, but e.g. ndt7 will need this feature.

The caller needs to do slightly more work because they will
need to process (or discard) the async events.

Yet, it makes sense to move in this direction because we
will want to return async events in mst nettest.