geolookup: don't ignore gzip.Close return value
While ignoring the error when closing a file open for read is
fine, doing the same for a gzip Reader is suboptimal because
we may miss stuff like checksumming error.

The documentation is a bit vague on this respect, but if I
understand the code correctly, we may reach EOF of the input
reader and then realize that the block is corrupt.